test(VBadge): update tests
